+++ relay-server/src/utils/multipart.rs
@@ -1,6 +1,7 @@
 use std::io;
 use std::task::Poll;
 
+use axum::extract::FromRequest;
 use axum::extract::Request;
 use bytes::{Bytes, BytesMut};
 use futures::{StreamExt, TryStreamExt};
@@ -9,6 +10,8 @@ use relay_config::Config;
 use serde::{Deserialize, Serialize};
 
 use crate::envelope::{AttachmentType, ContentType, Item, ItemType, Items};
+use crate::extractors::Remote;
+use crate::service::ServiceState;
 
 /// Type used for encoding string lengths.
 type Len = u32;
@@ -152,7 +155,7 @@ pub fn get_multipart_boundary(data: &[u8]) -> Option<&str> {
         .and_then(|slice| std::str::from_utf8(&slice[2..]).ok())
 }
 
-pub async fn multipart_items<F>(
+async fn multipart_items<F>(
     mut multipart: Multipart<'_>,
     mut infer_type: F,
     config: &Config,
@@ -163,6 +166,7 @@ where
 {
     let mut items = Items::new();
     let mut form_data = FormDataWriter::new();
+    let mut attachments_size = 0;
 
     while let Some(field) = multipart.next_field().await? {
         if let Some(file_name) = field.file_name() {
@@ -176,6 +180,14 @@ where
                 Err(multer::Error::FieldSizeExceeded { .. }) if ignore_large_fields => continue,
                 Err(err) => return Err(err),
                 Ok(bytes) => {
+                    attachments_size += bytes.len();
+
+                    if attachments_size > config.max_attachments_size() {
+                        return Err(multer::Error::StreamSizeExceeded {
+                            limit: config.max_attachments_size() as u64,
+                        });
+                    }
+
                     if let Some(content_type) = content_type {
                         item.set_payload(content_type.as_ref().into(), bytes);
                     } else {
@@ -277,10 +289,59 @@ impl futures::Stream for LimitedField<'_> {
     }
 }
 
-pub fn multipart_from_request(
-    request: Request,
-    config: &Config,
-) -> Result<Multipart<'static>, multer::Error> {
+/// Wrapper around [`multer::Multipart`] that checks each field is smaller than
+/// `max_attachment_size` and that the combined size of all fields is smaller than
+/// 'max_attachments_size'.
+pub struct ConstrainedMultipart(pub Multipart<'static>);
+
+impl FromRequest<ServiceState> for ConstrainedMultipart {
+    type Rejection = Remote<multer::Error>;
+
+    async fn from_request(
+        request: Request,
+        _state: &ServiceState,
+    ) -> Result<Self, Self::Rejection> {
+        multipart_from_request(request).map(Self).map_err(Remote)
+    }
+}
+
+impl ConstrainedMultipart {
+    pub async fn items<F>(self, infer_type: F, config: &Config) -> Result<Items, multer::Error>
+    where
+        F: FnMut(Option<&str>, &str) -> AttachmentType,
+    {
+        multipart_items(self.0, infer_type, config, false).await
+    }
+}
+
+/// Wrapper around [`multer::Multipart`] that skips over fields which are larger than
+/// `max_attachment_size`. These fields are also not taken into account when checking that the
+/// combined size of all fields is smaller than `max_attachments_size`.
+#[allow(dead_code)]
+pub struct UnconstrainedMultipart(pub Multipart<'static>);
+
+impl FromRequest<ServiceState> for UnconstrainedMultipart {
+    type Rejection = Remote<multer::Error>;
+
+    async fn from_request(
+        request: Request,
+        _state: &ServiceState,
+    ) -> Result<Self, Self::Rejection> {
+        multipart_from_request(request).map(Self).map_err(Remote)
+    }
+}
+
+#[cfg_attr(not(any(test, sentry)), expect(dead_code))]
+impl UnconstrainedMultipart {
+    pub async fn items<F>(self, infer_type: F, config: &Config) -> Result<Items, multer::Error>
+    where
+        F: FnMut(Option<&str>, &str) -> AttachmentType,
+    {
+        multipart_items(self.0, infer_type, config, true).await
+    }
+}
+
+pub fn multipart_from_request(request: Request) -> Result<Multipart<'static>, multer::Error> {
     let content_type = request
         .headers()
         .get("content-type")
@@ -288,13 +349,9 @@ pub fn multipart_from_request(
         .unwrap_or("");
     let boundary = multer::parse_boundary(content_type)?;
 
-    // Only enforce the stream limit here as the `per_field` limit is enforced by `LimitedField`.
-    let limits = multer::SizeLimit::new().whole_stream(config.max_attachments_size() as u64);
-
-    Ok(Multipart::with_constraints(
+    Ok(Multipart::new(
         request.into_body().into_data_stream(),
         boundary,
-        multer::Constraints::new().size_limit(limits),
     ))
 }
 
@@ -385,7 +442,7 @@ mod tests {
               ok\r\n\
               --X-BOUNDARY--\r\n";
 
-        let stream = futures::stream::once(async { Ok::<_, Infallible>(data) });
+        let stream = futures::stream::once(async move { Ok::<_, Infallible>(data) });
         let multipart = Multipart::new(stream, "X-BOUNDARY");
 
         let config = Config::from_json_value(serde_json::json!({
@@ -394,8 +451,9 @@ mod tests {
             }
         }))?;
 
-        let items =
-            multipart_items(multipart, |_, _| AttachmentType::Attachment, &config, true).await?;
+        let items = UnconstrainedMultipart(multipart)
+            .items(|_, _| AttachmentType::Attachment, &config)
+            .await?;
 
         // The large field is skipped so only the small one should make it through.
         assert_eq!(items.len(), 1);
@@ -420,23 +478,19 @@ mod tests {
               ok\r\n\
               --X-BOUNDARY--\r\n";
 
-        let stream = futures::stream::once(async { Ok::<_, Infallible>(data) });
+        let stream = futures::stream::once(async move { Ok::<_, Infallible>(data) });
 
         let config = Config::from_json_value(serde_json::json!({
             "limits": {
                 "max_attachments_size": 5
             }
         }))?;
-        let limits = multer::SizeLimit::new().whole_stream(config.max_attachments_size() as u64);
 
-        let multipart = Multipart::with_constraints(
-            stream,
-            "X-BOUNDARY",
-            multer::Constraints::new().size_limit(limits),
-        );
+        let multipart = Multipart::new(stream, "X-BOUNDARY");
 
-        let result =
-            multipart_items(multipart, |_, _| AttachmentType::Attachment, &config, true).await;
+        let result = UnconstrainedMultipart(multipart)
+            .items(|_, _| AttachmentType::Attachment, &config)
+            .await;
 
         // Should be warned if the overall stream limit is being breached.
         assert!(result.is_err_and(|x| matches!(x, multer::Error::StreamSizeExceeded { limit: _ })));